==> analysis/04_Catchments_Transport_SpatialNets.Rmd <==
---
title: "Settlement Persistence Project, SBOM Script #4:"
subtitle: "Catchment Territories, Transport and Spatial Network Models"
author: "Rudolf Cesaretti"
date: "Last run on `r Sys.Date()`"
output:
  html_document:
    toc: true
    df_print: paged
    number_sections: true
bibliography: References.bib
csl: apa.csl
link-citations: yes
---

```{css, echo=FALSE}
pre {
  max-height: 300px;
  overflow-y: auto;
}

pre[class] {
  max-height: 300px;
}
```

```{r, setup, include=FALSE,echo=FALSE, message=FALSE,warning=FALSE}
require(knitr)
# Set so that long lines in R will be wrapped:
opts_chunk$set(tidy.opts=list(width.cutoff=75),tidy=TRUE)
#
rm(list = ls())
```

I do five things in this R markdown document: 

  1. Create site catchment areas
  2. Calculate catchment area transport route densities
  3. Calculate catchment area demographic densities
  4. Define spatial network models for AggSites/Catchments and associated metrics
  5. Reorganize the data and export for Script #5
  
  
# Setup 

All of the data and scripts are downloadable from the [new ASU SettlementPersist2022 github repository](https://https://github.com/rcesaret/ASUSettlementPersist2022), which can be downloaded locally as a .zip folder or cloned to your own account.

Either way, once you have done so, you will need to modify the working directory (setwd("C:/...)") path and "dir" variables in the code chunk below to match the repository location on your computer.

```{r, label='Set Local Directory Location', message=FALSE,warning=FALSE}

wd <- list()

#SET YOUR LOCAL DIRECTORY LOCATION HERE:
wd$dir <- "C:/Users/rcesaret/Dropbox (ASU)/ASUSettlementPersist2022/"
#wd$dir <- "C:/Users/TJ McMote/Dropbox (ASU)/ASUSettlementPersist2022"

wd$analysis <- paste0(wd$dir,"analysis/")
wd$data_r <- paste0(wd$dir,"data-raw/")
wd$data_p <- paste0(wd$dir,"data-processed/")
wd$data_f <- paste0(wd$dir,"data-final-outputs/")
wd$figs <- paste0(wd$dir,"figures/")
wd$funcs <- paste0(wd$dir,"functions/")

```


## Load R Packages and Custom Functions

```{r, label='Load Libraries', message=FALSE,warning=FALSE}
# Package names
packages <- c("rgdal", "rgeos", "sp", "sf", "GISTools", "raster", "Matrix", 
              "gdistance", "lwgeom", "tidyverse","tidyr", "stars", "dismo", 
               "spatialEco", "whitebox", "classInt", "deldir", "spdep",
              "igraph", "dbscan", "cowplot", "deldir", "cccd", "ggraph", 
              "geosphere", "statnet", "intergraph","ggnewscale")
              
              #, "data.table", "zoo", "era", "JOPS", "mgcv","", , "ggrepel","ggridges", "movecost",  "datplot", "scales",

# Install packages not yet installed
installed_packages <- packages %in% rownames(installed.packages())
if (any(installed_packages == FALSE)) {
  install.packages(packages[!installed_packages])
}

# load packages
invisible(lapply(packages, library, character.only = TRUE))

rm(packages,installed_packages)

#Read in custom R functions located in the wd$funcs directory folder
FUNCS <- list("splitByAttributes.R", "CatchmentPolys.R", "SpatialNets.R", 
              "net_stats.R", "bbox_polygon.R")
invisible(lapply(FUNCS, function(x) source(paste0(wd$funcs,x))))
rm(FUNCS)

```


## Import Data

Data we are importing:

  1. The AggSite polygon data exported from script #2
  2. A simple polygon calculated in QGIS that specifies a hard outer border for the catchment areas of sites (constructed for sensitivity to survey borders and sites not included in the SBOM sample)
  3. Cost-distance matrices from script #3
  4. Least cost path rasters from script #3
  5. A raster hillshade basemap for the SBOM which includes the lakes

```{r, label='Import Data', message=FALSE,warning=FALSE}

#Agg Site polygon data
All_AggPoly <- readOGR(paste0(wd$data_p,"SBOM_AggPoly2.gpkg"))

#Catchment boundary limit polygon
CatchLims <- readOGR(paste0(wd$data_r,"CatchLims.gpkg"))
CatchLims.sf <- st_as_sf(CatchLims) #for ggplot

## Hillshade Basemap Raster with lake
HillshadeLake <- raster(paste0(wd$data_r, "HillshadeLake.tif"))
HillshadeLake <- rast(HillshadeLake, crs = 26914)
Hillshade.s <- st_as_stars(HillshadeLake) #for ggplot basemap

ord <- c("EF", "EF_MF", "MF", "MF_LF", "LF", "LF_TF", "TF", "TF_CL", "CL", "CL_ET", "ET", "ET_LTAzI", "LTAzI", "LTAzI_EA", "EA", "EA_LA", "LA")

#Cost-distance matrices
CD.mats <- readRDS(file=paste0(wd$data_p,"CDmats_list.RData"))

#Least cost path rasters
temp = list.files(path = paste0(wd$data_p,"lcp_dens_rasts/"), full.names = TRUE, pattern = ".grd")
nam.dens = list.files(path = paste0(wd$data_p,"lcp_dens_rasts/"), full.names = F, pattern = ".grd")
nam.dens = gsub('_lcpdens.grd', "", nam.dens)
lcp_dens_rasts = lapply(temp, raster)
names(lcp_dens_rasts) <- nam.dens
lcp_dens_rasts <- lcp_dens_rasts[ord]
names(lcp_dens_rasts) <- paste0(ord,"_lcpdens")

rm(temp, nam.distmat, nam.dens, ord) 
```


## Reorganize Settlement Data from Step #2

```{r, 'Reorganize Data from Step #2', message=FALSE,warning=FALSE}

# convert spatial polygons dataframe to spatial points dataframe
coor = All_AggPoly@data[,c("East","North")] #create separate dataframe of coordinates for spatial points dataframe
rownames(coor) <- as.numeric(rownames(coor)) #make sure rownames match
All_AggPts <- SpatialPointsDataFrame(coor, All_AggPoly@data, match.ID = TRUE) #convert to spatial points dataframe
proj4string(All_AggPts) <- CRS("+proj=utm +zone=14 +datum=NAD83 +units=m +no_defs") #make sure CRS matches polygons
All_AggPts = as(st_make_valid(st_as_sf(All_AggPts)), "Spatial") #make sure geometry is valid

#reorder spatial points dataframe by period
All_AggPts <- All_AggPts[order(All_AggPts$PeriodNum),]
All_AggPoly <- All_AggPoly[order(All_AggPoly$PeriodNum),]

# Split points by Phase, saved as list of spatial points dataframes
Pts_List <- splitByAttributes(spdata = All_AggPts, attr = "Period", suffix="_Pts") 
Poly_List <- splitByAttributes(spdata = All_AggPoly, attr = "Period", suffix="_Poly") 

# rename spatial points dataframes to include period numbers as prefix
#PeriodNums <- list("1","2","3","4","5","6","7","8","9","10","11","12","13","14","15","16","17")#include period number
#names(Pts_List) <- paste0(PeriodNums,"_",names(Pts_List))
#names(Poly_List) <- paste0(PeriodNums,"_",names(Poly_List))

rm(coor)# PeriodNums,
```


# Defining Catchment Areas

In order to characterize sites in terms of the characteristics of their immediate hinterlands, we can calculate their 'catchment areas.' These are GIS-calculated polygon representing the territorial hinterland of a settlement used for subsistence (etc). 

Here, we calculate settlement catchment areas by integrating two methods, implemented together using the custom "CatchmentPolys" function from 'SBOM_Script4_Functions.R'. 

The core method employs a modified version of 'Voronoi diagrams' (also known as 'Thiessen polygons'). Voronoi polygons partition a Euclidean plane into a polygon topology around a set of nodes with equidistant borders between nodes. However, instead of calculating Voronoi polygons from single nodes, the "CatchmentPolys" function uses the borders of site polygons. _To accomplish this, a Voronoi polygon is calculated for each point of a site's polygon, and then all of the Voronoi polygons belonging to each site are re-aggregated together into a single catchment polygon._ (The script also optionally allows the user to specify normal centroid/node-based Voronoi diagrams.) This method was implemented because of larger (usually urban) settlements in densely settled areas. In such cases, Voronoi polygons using centroid nodes frequently draw catchment borders on the immediate outskirts of sites (or even within the site's borders). Yet, the catchments of such sites should extend equidistant from the settlement border rather than their polygon centroid. As such, all else being equal, my method makes sure that sites with larger aerial extent will have larger catchment areas. In addition, these Voronoi polygons are restricted to the area within an outer border/limit of the SBOM region (constructed for sensitivity to survey borders and sites not included in the SBOM sample). This is a a simple polygon layer created in QGIS, used as an input to the "CatchmentPolys" function. 

Second, the catchment areas of sites are restricted to a 5 km radius around the centroid of the site polygon. This was done because cross-cultural research has demonstrated that traditional agriucltural settlements do not commonly intensively exploit areas beyond a 3-5 km / 1 hr travel time due to transport costs and commute time [see e.g. @Chisholm1979; @Roper1979; @Higgs1972; @Jarman1982]. To maximize speed of computation, I make the simplified approximation that 1 hr at pace ~= 3-5 km, although a cost-surface could instead be used (slowing computation). In periods where there are only a few settlements separated by large areas, this 5 km radius prevents the Voronoi polygons from encompassing vast swaths of territory. While these vast areas may well have been 'controlled' by such settlements, they do not accurately characterize the environmental/topographical context of their subsistence. (This 5 km threshold radius is irrelevant for most periods, where nearest neighbors are much closer than 5 km). Because the cost-distance radius is contextually variable, often ranging from from 3-5 km, I have calculated it at the higher 5 km level _from the centroid_ rather than from the edges of sites (as done for the Voronoi polygons). 

Below is an example of the "CatchmentPolys" function and its output:

```{r, 'Demonstrate Catchment Areas', message=FALSE,warning=FALSE}

ex1 <- Poly_List[[9]] #choose a period for which to calculate catchments

## Using the CatchmentPolys function

example <- CatchmentPolys(
              sitepolys = ex1, #input polygons from a single period
              border = CatchLims, #input polygon of the hard outer limit
              dist_lim = 5000, # set the radius distance threshold (in meters)
              method = "borders", # choice of method for Voronoi polygons: 
              #either "borders" or "centroids" (i.e. Voronoi polygons 
              #calculated from the borders of polygons or from their centroids)
              plot.results = F #whetherto output a plot of resulting catchments
              )


```
```{r, 'example plot of catchment areas', echo=FALSE, message=FALSE,warning=FALSE}

Sites.sf <- st_as_sf(Poly_List[[9]])
Catch.sf = st_as_sf(example)

ggp1 = ggplot() +  geom_stars(data = Hillshade.s)+
  scale_fill_gradientn(colours = c("turquoise", "black", "white"), 
               values = scales::rescale(c(-9999, -161, 254)), guide="none")+
  ggnewscale::new_scale_fill()+
  geom_sf(data = CatchLims.sf, aes(geometry = geometry), color="black", size=2, alpha = 0) +
  geom_sf(data = Catch.sf, aes(geometry = geometry),color="black", size=1, alpha = 0.25, fill="blue") +
  geom_sf(data = Sites.sf, aes(geometry = geometry),color="black", size=0.1, fill="red") +
  labs(title= "Defining Settlement Catchment Areas", subtitle = "Period 9, CL (c.AD 100-550)")+
  coord_sf()+
  theme_void()+
  theme(plot.title = element_text(hjust = 0.5, face="bold", size=16),
        plot.subtitle = element_text(hjust = 0.5, face="bold", size=14),
        plot.background = element_rect(fill = "white", colour = NA))

#save figure
ggsave("CatchmentAreaExample.png", plot = ggp1, device = "png", path = wd$figs, scale = 1, width = 7.5, height = 5,   units = "in",  dpi = 1800)

rm(example, ex1, ggp1, Sites.sf, Catch.sf)

# import figure for display
knitr::include_graphics(paste0(wd$figs,"CatchmentAreaExample.png"), FALSE)
```
## Create Catchment Areas for All Sites/Periods

Having explained the methods and demonstrated the use of the function, we can now calculate catchment areas for all periods by looping over them:

```{r, 'Calculate Catchment Areas for All Periods', message=FALSE,warning=FALSE}

# Define a vector of strings specifying method parameters.
# This was done because period 14 for some reason was not 
# behaving with the "borders" method. May need to debug 
# the script further...
# Regardless, the output for this period looks just fine 
# due to the lack of urban sites with large areas.
methods <- c("borders", "borders", "borders", "borders", "borders", "borders", "borders", "borders", "borders", "borders", "borders", "borders", "borders", "centroids", "borders", "borders", "borders")

catch_dist_lim <- 5000 # catchment areas limited to distance of X METERS from borders of site polygon

Catch_List <- list() #create list to store outputs

for (i in 1:length(Poly_List)){
  
  xx = as.character(methods[i]) 
  
  catch <- CatchmentPolys(sitepolys = Poly_List[[i]], 
                          border = CatchLims, 
                          dist_lim = catch_dist_lim,
                          method = xx,
                          plot.results = F)
  
  Catch_List[[i]] <- catch
  
}

#rename output list objects
sitepoly.names <- names(Poly_List) 
sitepoly.names <- gsub('_Poly', "_SitePoly", sitepoly.names)
names(Poly_List) <- sitepoly.names
catch.names <- sitepoly.names
catch.names <- gsub('_SitePoly', "_CatchPoly", catch.names)
names(Catch_List) <- catch.names

rm(methods, catch_dist_lim, xx, catch)
```



# Catchment Area Transport Route Density

Here we calculate the "transport route density" variable for site catchment areas from the least cost path rasters calculated in script #3. This variable gives us a proxy for how central a settlement's territory may have been to the major transport arteries of a given period. Below is an example map of the LCP density rasters.

```{r, 'example plot of LCP Density', echo=FALSE, message=FALSE,warning=FALSE}

Sites.sf <- st_as_sf(Poly_List[[9]])
Catch.sf = st_as_sf(Catch_List[[9]])
lcp_dens.s <- st_as_stars(lcp_dens_rasts[[9]])

ggp1 = ggplot() +  geom_stars(data = Hillshade.s)+
  scale_fill_gradientn(colours = c("turquoise", "black", "white"), 
               values = scales::rescale(c(-9999, -161, 254)), guide="none")+
  ggnewscale::new_scale_fill()+
  geom_stars(data = lcp_dens.s, alpha=0.5)+
  scale_fill_gradientn(colours = c(NA, "black", "black"), 
               values = scales::rescale(c(0, 50, 273)), guide="none")+
  geom_sf(data = CatchLims.sf, aes(geometry = geometry), color="black", size=2, alpha = 0) +
  geom_sf(data = Catch.sf, aes(geometry = geometry),color="black", size=0.35, alpha = 0) +
  geom_sf(data = Sites.sf, aes(geometry = geometry),color="black", size=0.1, fill="red") +
  labs(title= "Least Cost Path Density", subtitle = "Period 9, CL (c.AD 100-550)")+
  coord_sf()+
  theme_void()+
  theme(plot.title = element_text(hjust = 0.5, face="bold", size=16),
        plot.subtitle = element_text(hjust = 0.5, face="bold", size=14),
        plot.background = element_rect(fill = "white", colour = NA))

#save figure
ggsave("LCPDensityExample.png", plot = ggp1, device = "png", path = wd$figs, scale = 1, width = 7.5, height = 5,   units = "in",  dpi = 1800)

rm(lcp_dens.s, ggp1, Sites.sf, Catch.sf)

# import figure for display
knitr::include_graphics(paste0(wd$figs,"LCPDensityExample.png"), FALSE)
```


Transport route density is calculated _in two ways for each period_:

  1. as a **percentage of the total** least cost path area (90m x 90m raster cells) in the SBOM
  2. as an **ordinal rank** of the above percentages among all site catchments (highest % = 1)

Here, again, we will loop over all periods using the LCP rasters imported from script #3:


```{r, 'Catchment Area Transport Route Density', message=FALSE,warning=FALSE}

Catch_List2 <- list() #create list for outputs

for(i in 1:length(Catch_List)){
  
  #define catchment area polys and LCP rasters from lists as temp objects
  
  tmp.p <- Catch_List[[i]] 
  tmp.p2 <- vect(tmp.p)
  tmp.r <- rast(lcp_dens_rasts[[i]])
  
  #calculate sum/total LCP area of raster
  sum.lyr <- as.numeric(terra::global(tmp.r, fun='sum', na.rm=TRUE))
  
  #calculate transport route density as percentage
  TranspDens.pct <- terra::extract(tmp.r, tmp.p2, fun=sum)#
  tmp.p@data$TranspDens.pct <- TranspDens.pct[,2]/sum.lyr * 100
  
  #calculate transport route density as rank
  tmp.p@data$TranspDens.rank <- rank(-tmp.p@data$TranspDens.pct)
  
  Catch_List2[[i]] <- tmp.p #save outputs to list
}

#rename catchment list and its objects accordingly
names(Catch_List2) <- catch.names
Catch_List <- Catch_List2

rm(Catch_List2)

```


# Catchment Area Demographic Densities

Here we calculate catchment population densities -- population divided by catchment area (ha) -- to provide a proxy for the intensity of land use. In addition to the raw values, relative global values are also calculated. These include the site/catchment's value (A) as a proportion of the global maximum catchment pop density, and (B) as a global rank of catchment pop density values. 

These three metrics are also calculated for the catchment area (ha) beyonjd the site polygon limits (i.e. total catchment area minus site area). This provides an idea of specifically rural land use intensity.

Altogether, the metrics are as follows:

  1. **Catch_Popdens** = Catchment Population Density
  2. **Catch_Popdens_PropMax** = Catchment Population Density as a proportion of the global maximum
  3. **Catch_Popdens_Rank** = Catchment Population Density global rank
  4. **CatchB_Popdens** = Catchment Population Density for the area beyond the site polygon
  5. **CatchB_Popdens_PropMax** = Catchment Population Density for the area beyond the site polygon as a proportion of the global maximum 
  6. **CatchB_Popdens_Ran** = Global rank of Catchment Population Density for the area beyond the site polygon


```{r, 'Catchment Area Demographic Densities', message=FALSE,warning=FALSE}

Catch_List2 <- list() #create output list

for(i in 1:length(Catch_List)){
  
  tmp.p <- Catch_List[[i]] #define catchment area polys as temp object

  tmp.p@data$Catch_Popdens <- tmp.p@data$Population.s2/tmp.p@data$Catchment_ha
  
  tmp.p@data$CatchB_Popdens <- tmp.p@data$Population.s2/tmp.p@data$CatchmentBeyond_ha
  
  tmp.p@data$Catch_Popdens_PropMax <- tmp.p@data$Catch_Popdens/(max(tmp.p@data$Catch_Popdens))
  
  tmp.p@data$Catch_Popdens_Rank <- rank(-tmp.p@data$Catch_Popdens, na.last = "keep", ties.method = "average")
  
  tmp.p@data$CatchB_Popdens_PropMax <- (tmp.p@data$CatchB_Popdens / (max(tmp.p@data$CatchB_Popdens)))
  
  tmp.p@data$CatchB_Popdens_Rank <- rank(-tmp.p@data$CatchB_Popdens, na.last = "keep", ties.method = "average")
  
  tmp.p@data <- tmp.p@data %>% rowwise() %>% 
    mutate(
        Catch_Popdens_Rank = ifelse(is.na(Catch_Popdens_Rank), max(Catch_Popdens_Rank)+1, Catch_Popdens_Rank),
        CatchB_Popdens_Rank = ifelse(is.na(CatchB_Popdens_Rank), max(CatchB_Popdens_Rank)+1, CatchB_Popdens_Rank)) %>% ungroup()
  
  Catch_List2[[i]] <- tmp.p #save to output list
}

# rename list items
names(Catch_List2) <- catch.names
Catch_List <- Catch_List2#

rm(Catch_List2)
```


# Modelling Transport Networks

Spatial network models provide one way of modeling the transportation networks interconnecting the settlements of our survey regions. The approach taken here employs the methods and code from [Chapter 7 of Peeples and Brughmans' Online Companion to Archaeological Network Science](https://book.archnetworks.net/spatialnetworks) [see @Peeples_Brughmans_2022; @Brughmans_Peeples_2022], as well as from the [spdep R package and its vignettes and author publications](https://cran.r-project.org/web/packages/spdep/) [@Bivand2018; @Bivand2022; @Bivand2013]. A combination of the following spatial network methods is to model potential transport networks:

  * **Gabriel Graphs** -- Edges are formed between pairs of nodes if there are no other nodes located within a circle whose diameter is a straight line between those two nodes (see below)

```{r, echo=FALSE, fig.dim = c(4, 2), fig.cap="Gabriel Graphs"}
knitr::include_graphics(paste0(wd$figs,"GabrielGraph.PNG"), FALSE)
```
  
  * **Relative Neighborhood Networks (RNN)** -- Very similar to the Gabriel graph, edges are formed between pairs of nodes if there are no other nodes located within the 'Venn Diagram' intersection of two circles surrounding the two nodes. As seen below, the radii of these circles is the the distance between the two nodes
  
```{r, echo=FALSE, fig.dim = c(3, 2), fig.cap="Relative Neighbor Network"}
knitr::include_graphics(paste0(wd$figs,"RelativeNeighborNet.PNG"), FALSE)
```
  
  * **Sphere of Influence Graph (SOI)** -- Similar again to both Gabrial graphs and relative neighbor networks, sphere of inflence graphs have nodes construct a single circle from the distance to their nearest neighbor, and then form edges with all other nodes whose circles intersect their circle (see below)
  
```{r, echo=FALSE, fig.dim = c(3, 2), fig.cap="Delaunay Triangulation"}
knitr::include_graphics(paste0(wd$figs,"SOIgraph.PNG"), FALSE)
```
  
  * **Delaunay Triangulation** -- Network topology formed among nodes whose corresponding Voronoi/Thessien polygons share a border (see below)
  
```{r, echo=FALSE, fig.dim = c(3, 2), fig.cap="Delaunay Triangulation"}
knitr::include_graphics(paste0(wd$figs,"DelaunayTriangulation.PNG"), FALSE)
```
  
  * **Polygon Border Neighbors** -- A more generalized form of Delaunay triangulation, this method simply uses any set of polygons and creates a graph topology from all polygons that share an edge. _Using this method, we can construct spatial networks from our quasi-Voronoi catchment areas calculated above._
  * **K Nearest Neighbors (KNN)** -- nodes form edges with their _k_ nearest neighbors (where _k_ is a parameter specified exogenously)
  * **Maximum Distance Networks** -- each node forms edges with all other nodes that fall within a user-specified distance/radius of that node

Examples of each of these networks in the SBOM are shown in the figure below:

```{r, 'transport net figs calc', echo=FALSE, warning=FALSE, message=FALSE}
p = 9 # Select Period Number

ggbasemap <- ggplot() +  geom_stars(data = Hillshade.s)+
  scale_fill_gradientn(colours = c("turquoise", "black", "white"), 
                       values = scales::rescale(c(-9999, -161, 254)), guide="none")+
  xlim(485000, 529500)+ylim(2106800, 2150600)+theme_void()


spnet <- SpatialNets(CD.mat = CD.mats[[p]]/3600,#select cost-dist matrix and convert to hrs (hrs=seconds/3600)
                        nodes = Pts_List[[p]]@data,#create node data
                        coords = coordinates(Pts_List[[p]]),# coordinate pairs for sites 
                        Pts.sf = st_as_sf(Pts_List[[p]]),# sf class site points
                        Sites.sf = st_as_sf(Poly_List[[p]]),# sf class site polygons
                        Catch.sf = st_as_sf(Catch_List[[p]]),# sf class catchment polygons
                        limits = CatchLims, #
                        bounds = extent(limits),
                        window = c(bounds@xmin, bounds@xmax, bounds@ymin, bounds@ymax),
                        ggplots = "MultiPlot",
                        ggbasemap = ggbasemap, 
                        mode = "multi",
                        method = c("gabriel", "rnn", "delaunay", "soi", "poly", "knn", "maxdist"),
                        combine = "weighted",
                        plotNodeStat = "centr_clos_n",
                        ggPlotTitle = "SBOM Spatial Networks: CL (AD 100-550)",
                        knn_k = 5,
                        maxdist = 1,
                        outputs = "none")

ggsave("MultiSpatNetModels_CL.png", plot = spnet[[1]], device = "png", path = wd$figs, scale = 1, width = 6.5, height = 6.5,   units = "in",  dpi = 2000)

p = 13 # Select Period Number

spnet <- SpatialNets(CD.mat = CD.mats[[p]]/3600,#select cost-dist matrix and convert to hrs (hrs=seconds/3600)
                        nodes = Pts_List[[p]]@data,#create node data
                        coords = coordinates(Pts_List[[p]]),# coordinate pairs for sites 
                        Pts.sf = st_as_sf(Pts_List[[p]]),# sf class site points
                        Sites.sf = st_as_sf(Poly_List[[p]]),# sf class site polygons
                        Catch.sf = st_as_sf(Catch_List[[p]]),# sf class catchment polygons
                        limits = CatchLims, #
                        bounds = extent(limits),
                        window = c(bounds@xmin, bounds@xmax, bounds@ymin, bounds@ymax),
                        ggplots = "MultiPlot",#,"PlotResults", "none"),
                        ggbasemap = ggbasemap, 
                        mode = "multi",#c("single", "multi", "multicombine"),
                        method = c("gabriel", "rnn", "delaunay", "soi", "poly", "knn", "maxdist"),
                        combine = "weighted",#("weighted","unweighted"),
                        plotNodeStat = "centr_clos_n",
                        ggPlotTitle = "SBOM Spatial Networks: LTAzI (950-1200)",
                        knn_k = 5,
                        maxdist = 1,
                        outputs = "none")

ggsave("MultiSpatNetModels_LTAzI.png", plot = spnet[[1]], device = "png", path = wd$figs, scale = 1, width = 6.5, height = 6.5,   units = "in",  dpi = 2000)

p = 9 # Select Period Number

spnet <- SpatialNets(CD.mat = CD.mats[[p]]/3600,#select cost-dist matrix and convert to hrs (hrs=seconds/3600)
                        nodes = Pts_List[[p]]@data,#create node data
                        coords = coordinates(Pts_List[[p]]),# coordinate pairs for sites 
                        Pts.sf = st_as_sf(Pts_List[[p]]),# sf class site points
                        Sites.sf = st_as_sf(Poly_List[[p]]),# sf class site polygons
                        Catch.sf = st_as_sf(Catch_List[[p]]),# sf class catchment polygons
                        limits = CatchLims, #
                        bounds = extent(limits),
                        window = c(bounds@xmin, bounds@xmax, bounds@ymin, bounds@ymax),
                        ggplots = "PlotResults",#,"PlotResults", "none"),
                        ggbasemap = ggbasemap, 
                        mode = "single",#c("single", "multi", "multicombine"),
                        method = "gabriel",
                        combine = "weighted",#("weighted","unweighted"),
                        plotNodeStat = "centr_clos_n",
                        ggPlotTitle = "SBOM CL (AD 100-550) Transport Network Modelled as",
                        knn_k = 5,
                        maxdist = 1,
                        outputs = "none")

ggsave("GabrielNet_CL.png", plot = spnet[[1]], device = "png", path = wd$figs, scale = 1, width = 6.5, height = 6,   units = "in",  dpi = 2000)
```

Gabriel, RNN and KNN-5 perform the best

```{r, echo=FALSE, fig.cap="Classic Period Spatial Network Models", out.width = '100%'}
knitr::include_graphics(paste0(wd$figs,"MultiSpatNetModels_CL.png"), FALSE)
```


```{r, echo=FALSE, fig.cap="LT-AzI Period Spatial Network Models", out.width = '100%'}
knitr::include_graphics(paste0(wd$figs,"MultiSpatNetModels_LTAzI.png"), FALSE)
```


```{r, echo=FALSE, fig.cap="Final Proposed Model for Transport Nets: Classic Period", out.width = '100%'}
knitr::include_graphics(paste0(wd$figs,"GabrielNet_CL.png"), FALSE)
```


# Calculate Networks for All Periods 


```{r, 'calulate transport nets for all periods', warning=FALSE, message=FALSE}
per <- c("EF", "EF_MF", "MF", "MF_LF", "LF", "LF_TF", "TF", "TF_CL", "CL", "CL_ET", "ET", "ET_LTAzI", "LTAzI", "LTAzI_EA", "EA", "EA_LA", "LA")

TrsprtNet_CDmatList <- list()
TrsprtNet_igList <- list()
TrsprtNet_igwtList <- list()
TrsprtNet_lines_sfList <- list()
TrsprtNet_adjmat <- list()
TrsprtNet_adjmatCD <- list()
TrsprtNet_nb <- list()
TrsprtNet_ggplots <- list()

for (p in 1:length(Pts_List)){
  
  tmp <- SpatialNets(CD.mat = CD.mats[[p]]/3600,
                        nodes = Pts_List[[p]]@data,
                        coords = coordinates(Pts_List[[p]]),
                        Pts.sf = st_as_sf(Pts_List[[p]]),
                        Sites.sf = st_as_sf(Poly_List[[p]]),
                        Catch.sf = st_as_sf(Catch_List[[p]]),
                        limits = CatchLims, #
                        bounds = extent(limits),
                        window = c(bounds@xmin, bounds@xmax, bounds@ymin, bounds@ymax),
                        ggplots = "PlotResults",#,"PlotResults", "none"),
                        ggbasemap = ggbasemap, 
                        mode = "single",#c("single", "multi", "multicombine"),
                        method = "gabriel",
                        combine = "weighted",#("weighted","unweighted"),
                        plotNodeStat = "centr_clos_n",
                        ggPlotTitle = per[p],
                        knn_k = 5,
                        maxdist = 1,
                        outputs = c("net_distmat", "netstats_df", "igraph", 
                                    "igraph_wt", "lines_sf", "adjmat", "adjmatCD",
                                    "nb"))
  
  TrsprtNet_CDmatList[[p]] <- tmp[[1]][[1]]
  write.csv(tmp[[1]][[1]], file=paste0(wd$data_p,"TrsprtNet_CDmats/",per[p],"_TrsprtNet_CDmat.csv"))
  stat <- tmp[[2]][[1]][,c(5:ncol(tmp[[2]][[1]]))]
  Poly_List[[p]] <- cbind(Poly_List[[p]],stat)
  TrsprtNet_igList[[p]] <- tmp[[3]][[1]]
  TrsprtNet_igwtList[[p]] <- tmp[[4]][[1]]
  TrsprtNet_lines_sfList[[p]] <- tmp[[5]][[1]]
  TrsprtNet_adjmat[[p]] <- tmp[[6]][[1]]
  TrsprtNet_adjmatCD[[p]] <- tmp[[7]][[1]]
  TrsprtNet_nb[[p]] <- tmp[[8]][[1]]
  ggtmp <- tmp[[9]]+scale_size(range = c(1,2))+theme(legend.position="none", plot.title = element_blank(),
                                          plot.subtitle = element_blank(),plot.margin = unit(c(0,0,0,0), "cm"))
  TrsprtNet_ggplots[[p]] <- ggtmp
}
names(TrsprtNet_CDmatList) <- per
names(TrsprtNet_igList) <- per
names(TrsprtNet_igwtList) <- per
names(TrsprtNet_lines_sfList) <- per
names(TrsprtNet_adjmat) <- per
names(TrsprtNet_adjmatCD) <- per
names(TrsprtNet_nb) <- per
names(TrsprtNet_ggplots) <- per

saveRDS(TrsprtNet_CDmatList, file=paste0(wd$data_p,"TrsprtNet_CDmatList.RData"))
saveRDS(TrsprtNet_igList, file=paste0(wd$data_p,"TrsprtNet_igList.RData"))
saveRDS(TrsprtNet_igwtList, file=paste0(wd$data_p,"TrsprtNet_igwtList.RData"))
saveRDS(TrsprtNet_lines_sfList, file=paste0(wd$data_p,"TrsprtNet_lines_sfList.RData"))
saveRDS(TrsprtNet_adjmat, file=paste0(wd$data_p,"TrsprtNet_adjmat.RData"))
saveRDS(TrsprtNet_adjmatCD, file=paste0(wd$data_p,"TrsprtNet_adjmatCD.RData"))
saveRDS(TrsprtNet_nb, file=paste0(wd$data_p,"TrsprtNet_nb.RData"))

ggplotgrid = plot_grid(TrsprtNet_ggplots[[1]], TrsprtNet_ggplots[[2]], TrsprtNet_ggplots[[3]],
                           TrsprtNet_ggplots[[4]], TrsprtNet_ggplots[[5]], TrsprtNet_ggplots[[6]],
                           TrsprtNet_ggplots[[7]], TrsprtNet_ggplots[[8]], TrsprtNet_ggplots[[9]],
                           labels = per[1:9], align="hv", nrow = 3, ncol = 3)
    
    title <- ggdraw() + draw_label("SBOM Gabriel Graph Transport Nets",fontface = 'bold',hjust = 0.5, size=20 )
    
    ggout <- plot_grid(title, ggplotgrid, ncol = 1, rel_heights = c(0.05,0.95)) + 
      theme(plot.background = element_rect(fill = "white", colour = NA))
    ggsave("GabrielTrnspNets_1.png", plot = ggout, device = "png", path = wd$figs, 
           scale = 1, width = 7, height = 6.5,   units = "in",  dpi = 2000)
      
    ggplotgrid = plot_grid(TrsprtNet_ggplots[[10]], TrsprtNet_ggplots[[11]], TrsprtNet_ggplots[[12]],
                           TrsprtNet_ggplots[[13]], TrsprtNet_ggplots[[14]], TrsprtNet_ggplots[[15]],
                           TrsprtNet_ggplots[[16]], TrsprtNet_ggplots[[17]], 
                           labels = per[10:17], align="hv", nrow = 3, ncol = 3)
    
    ggout <- plot_grid(title, ggplotgrid, ncol = 1, rel_heights = c(0.05,0.95)) + 
      theme(plot.background = element_rect(fill = "white", colour = NA))
    ggsave("GabrielTrnspNets_2.png", plot = ggout, device = "png", path = wd$figs, 
           scale = 1, width = 7, height = 6.5,   units = "in",  dpi = 2000)
```

```{r, echo=FALSE, fig.cap="SBOM Gabriel Graph Transport Nets, 1-9", out.width = '100%'}
knitr::include_graphics(paste0(wd$figs,"GabrielTrnspNets_1.png"), FALSE)
```

```{r, echo=FALSE, fig.cap="SBOM Gabriel Graph Transport Nets, 10-17", out.width = '100%'}
knitr::include_graphics(paste0(wd$figs,"GabrielTrnspNets_2.png"), FALSE)
```

```{r}
xx <-  do.call(rbind, Poly_List)
per_meanz <- xx@data %>% group_by(PeriodNum,Period) %>% summarise_at(vars(Area_ha:PopFwCont,FwOvlp.Area:FwOvlp.Pop,BwOvlp.Area:Persist,OccuTime:UrbOccuInertia,centr_deg:cntrlz_avg), mean, na.rm = TRUE)
per_sumz <- xx@data %>% group_by(PeriodNum,Period) %>% summarise_at(vars(Area_ha:PopFwCont,FwOvlp.Area:FwOvlp.Pop,BwOvlp.Area:Persist,OccuTime:UrbOccuInertia,centr_deg:cntrlz_avg), sum, na.rm = TRUE)
ggplot(per_meanz, aes(x=PeriodNum))+geom_line(aes(y=scales::rescale(cntrlz_avg)),color="blue")+geom_line(aes(y=scales::rescale(density)),color="black")+geom_line(aes(y=scales::rescale(trans_glob)),color="red")+geom_line(aes(y=scales::rescale(connectiv)),color="green3")
per_sumz$connectiv
```








# Recombining and Reorganizing the Data

```{r, 'Recombining the Data Part 2', message=FALSE, warning=FALSE}
# Convert lists of period-wise sites/catchments to single SPDF objects
All_Agg_SitePoly <-  do.call(rbind, Poly_List)
All_Agg_CatchPoly <- do.call(rbind, Catch_List)

# variables from site data that needs transfer over to catchment areas
colz1 = setdiff(colnames(All_Agg_SitePoly@data),colnames(All_Agg_CatchPoly@data))
# variables from catchment areas that needs transfer over to site data
colz2 = setdiff(colnames(All_Agg_CatchPoly@data),colnames(All_Agg_SitePoly@data))

#reorder the data to match
All_Agg_SitePoly <- All_Agg_SitePoly[order(All_Agg_SitePoly$AggSite),]
All_Agg_CatchPoly <- All_Agg_CatchPoly[order(All_Agg_CatchPoly$AggSite),]

#check to see that the two datasets are in the right order
#identical(All_Agg_SitePoly@data$AggSite, All_Agg_CatchPoly@data$AggSite)

# Site data to catchment areas
Site_to_Catch <- All_Agg_SitePoly@data %>% dplyr::select(!!!syms(colz1))
All_Agg_CatchPoly@data <- cbind(All_Agg_CatchPoly@data,Site_to_Catch)

# catchment area data to sites
Catch_to_Site <- All_Agg_CatchPoly@data %>% dplyr::select(!!!syms(colz2))
All_Agg_SitePoly@data <- cbind(All_Agg_SitePoly@data,Catch_to_Site)

#Reorganize the data
      
ordering <- c(
  #ID VARIABLES
      "AggSite","AggID","Site","East","North","SurvReg","Number","CerPhase","Period", 
      "PeriodType","PeriodLength","PeriodNum","PeriodBegin","PeriodEnd", 
      "OccSeqLoc","OccSeqLoc.Sites","SubOccSeqLoc","SubOccSeqLoc.Sites",
      "ComponentNum", "ComponentSites",
  #CHRONOLOGICAL VARIABLES
      "PeriodInterval", "PeriodBegin", "PeriodBegin.era", "PeriodMidpoint", 
      "PeriodMidpoint.era", "PeriodEnd", "PeriodEnd.era", "PeriodLength",
  #OCCUPATION VARIABLES (COUNTS)
      "Occ.EF","Occ.EF_MF","Occ.MF","Occ.MF_LF","Occ.LF","Occ.LF_TF",
      "Occ.TF","Occ.TF_CL","Occ.CL","Occ.CL_ET","Occ.ET","Occ.ET_LTAzI", 
      "Occ.LTAzI","Occ.LTAzI_EA","Occ.EA","Occ.EA_LA","Occ.LA","Occ.TOT",
  #SUBOCCUPATION VARIABLES (COUNTS)
      "SubOcc.EF","SubOcc.EF_MF","SubOcc.MF","SubOcc.MF_LF","SubOcc.LF",
      "SubOcc.LF_TF","SubOcc.TF","SubOcc.TF_CL","SubOcc.CL","SubOcc.CL_ET",
      "SubOcc.ET","SubOcc.ET_LTAzI","SubOcc.LTAzI","SubOcc.LTAzI_EA",
      "SubOcc.EA","SubOcc.EA_LA","SubOcc.LA","SubOcc.TOT",
  #SITE AREA AND OCCUPATIONAL DENSITY VARS
      "Area_ha","Perim_m2","SherdDens","Tot.Assemb","FwOvlp.Assemb", 
      "BwOvlp.Assemb", "Net.Assemb",
  #STEP #2 DEMOGRAPHIC VARIABLES
      "Population.s2","Log_Population.s2", "ApportAssemb", "PopDens.s2",
      "UrbanScale.s2", "UrbanPop.s2","RuralPop.s2", "PctUrban.s2","PctRural.s2",
      "Prior", "Observed", "MeanOccuProb", 
  #STEP #2 DEMOGRAPHIC RATES
      "Pct_deltaPop12", "Pct_deltaPop01", "r12_Pert", "r01_Pert",
      "r23_Pert", "rPert", "rPert0", "rPert2", 
      "Pct_UrbdeltaPop12", "Pct_UrbdeltaPop01", "Urb_r12_Pert", "Urb_r01_Pert",
      "Urb_r23_Pert", "Urb_rPert", "Urb_rPert0", "Urb_rPert2",
      "Abs_rPert","RS_rPert","LogRS_rPert","LogAbs_rPert","Q.rPert","Q.Abs_rPert",
      "Q.RS_rPert","Q.LogRS_rPert","Q.LogAbs_rPert","Urb_Abs_rPert",
      "Urb_RS_rPert","Urb_LogRS_rPert","Urb_LogAbs_rPert","Q.Urb_rPert","Q.Urb_Abs_rPert",
      "Q.Urb_RS_rPert","Q.Urb_LogRS_rPert","Q.Urb_LogAbs_rPert",
      "Qp.rPert","Qp.Abs_rPert","Qp.RS_rPert","Qp.LogRS_rPert","Qp.LogAbs_rPert",
  #STEP #1 DEMOGRAPHIC VARIABLES
      "Population.s1","PopDens.s1","UrbanScale.s1","UrbanPop.s1","RuralPop.s1", 
      "PctUrban.s1","PctRural.s1",
  #CONTINUITY VARIABLES
      "AreaBwCont","AreaFwCont","PopBwCont","PopFwCont","FwOvlp.Sites",
      "FwOvlp.Area","FwOvlp.Pop","BwOvlp.Sites","BwOvlp.Area","BwOvlp.Pop",
  #PERSISTENCE VARIABLES
      "Found","FoundInit","Abandon","Persist","DewarType",
      "OccuTime","OccuInertia","UrbOccuTime","UrbOccuInertia",
  #STEP #4 TRANSPORT NETWORK VARIABLES
      "TranspDens.pct", "TranspDens.rank","centr_deg","centr_btw","centr_eig",
      "centr_clos","centr_hrmo","centr_hub","centr_auth","centr_pgrk",
      "centr_deg_n","centr_btw_n","centr_eig_n","centr_clos_n","centr_hrmo_n",
      "centr_hub_n","centr_auth_n","centr_avg","trans_loc","trans_locavg",
      "density","connectiv","trans_glob","cntrlz_deg","cntrlz_btw","cntrlz_eig",
      "cntrlz_clo","cntrlz_hub","cntrlz_auth","cntrlz_deg_n","cntrlz_btw_n",
      "cntrlz_eig_n","cntrlz_clo_n","cntrlz_hub_n","cntrlz_auth_n","cntrlz_avg",
  #STEP #4 CATCHMENT AREA AND POP DENSITY VARIABLES
      "Catchment_ha", "CatchmentBeyond_ha", "Catch_Popdens", 
      "Catch_Popdens_PropMax","Catch_Popdens_Rank","CatchB_Popdens",
      "CatchB_Popdens_PropMax", "CatchB_Popdens_Rank", 
  #SURVEY METADATA
      "M_Sites","M_SiteCode","M_SiteName","M_FieldSite.Region",
      "M_FieldSite.Period","M_SurveyYearNumber","M_Supervisor","M_Map",
  #OLD tDAR BOM SURVEY VARIABLES
      "O_Elev","O_ElevMed","O_ElevMin","O_ElevMax","O_EZcode",
      "O_EnvironmentalZone","O_Soil","O_SoilMed","O_SoilMin","O_SoilMax",
      "O_Erosion","O_ErosionMed","O_ErosionMin","O_ErosionMax","O_ModernUse",
      "O_ModernSettlement","O_Rainfall","O_Area","O_MoundDomestic",
      "O_MoundCeremonial","O_MoundQuestionable","O_MoundTotal",
      "O_MoundRecorded","O_DMoundArea","O_Architecture","O_TerraceConfidence",
      "O_TerraceExtent","O_Sherd","O_SherdMed","O_SherdMin","O_SherdMax",
      "O_Rubble","O_RubbleMed","O_RubbleMin","O_RubbleMax","O_Population",
      "O_PopMin","O_PopMax","O_PopMethod","O_stcode","O_SiteType",
      "O_SubPeriod1","O_SubPeriod2","O_OccEF","O_OccMF","O_OccLF","O_OccTF",
      "O_OccCL","O_OccEC","O_OccMC","O_OccLC","O_OccET","O_OccLT","O_OccAZ",
      "O_OccEA","O_OccLA","O_OccTot","O_OccSeqLoc","O_SubOc1","O_SubOc2",
      "O_PdDupSite","O_Group","O_Comments") 

#Make sure everything is kosher
#setdiff(colnames(All_Agg_CatchPoly@data),ordering)
#setdiff(colnames(All_Agg_SitePoly@data),ordering)
#setdiff(ordering,colnames(All_Agg_CatchPoly@data))
#setdiff(ordering,colnames(All_Agg_SitePoly@data))

# Reorder the data for both sites and catchment areas
All_Agg_SitePoly@data <- All_Agg_SitePoly@data %>% dplyr::select(!!!syms(ordering))
All_Agg_CatchPoly@data <- All_Agg_CatchPoly@data %>% dplyr::select(!!!syms(ordering))
```


# Export Data for Script #5

```{r, 'Export Data for Step #4', message=FALSE,warning=FALSE}

#AggSite polygons
writeOGR(All_Agg_SitePoly, paste0(wd$data_p,"SBOM_AggSitePoly4.gpkg"), "SBOM_AggSitePoly4", driver = "GPKG", overwrite_layer=TRUE)

#Catchment areas
writeOGR(All_Agg_CatchPoly, paste0(wd$data_p,"SBOM_CatchPoly4.gpkg"), "SBOM_CatchPoly4", driver = "GPKG", overwrite_layer=TRUE)

```


# References
